Move getPriceTierForUnits to calypso-products package
